Copy pasta to new location
